db on delete cascade, auth for delete and patch

+++ routers/userRoutes.js
@@ -11,17 +11,17 @@ const router = Router();
 router.post("/login", (req, res) => {
   const { email, password } = req.body;
   if (!email || !password) {
-    return res.status(400).send("Invalid credetntials ");
+    return res.status(400).send("Missing some data! ");
   }
   const user = User.getUsersByEmail(email);
   if (!user) {
-    return res.status(404).send("User Not found");
+    return res.status(401).send("Unauthorized");
   }
   if (!bcrypt.compareSync(password, user.password)) {
-    return res.status(400).send("Invalid credetntials ");
+    return res.status(401).send("Unauthorized");
   }
   const token = jwt.sign({ id: user.id, email: user.email }, "secret_key", {
-    expiresIn: "30s",
+    expiresIn: "30m",
   });
   res.send({ token: token });
 });
@@ -29,9 +29,9 @@ router.post("/login", (req, res) => {
 router.patch("/:id", auth, async (req, res) => {
   const id = +req.params.id;
   if (id != req.userId) {
-    return res.status(400).send("Invalid user id");
+    return res.status(401).send("Unauthorized");
   }
-  const { name, email, password } = req.body;
+  const { email, password } = req.body;
   let user = User.getUsersById(id);
   let hashedPwd;
   if (password) {
@@ -39,12 +39,7 @@ router.patch("/:id", auth, async (req, res) => {
     hashedPwd = await bcrypt.hash(password, salt);
   }
 
-  User.updateUser(
-    id,
-    name || user.name,
-    email || user.email,
-    hashedPwd || user.password
-  );
+  User.updateUser(id, email || user.email, hashedPwd || user.password);
 
   user = User.getUsersById(id);
   delete user.password;
@@ -66,8 +61,8 @@ router.get("/me", auth, (req, res) => {
 });
 
 router.post("/register", async (req, res) => {
-  const { name, email, password } = req.body;
-  if (!name || !email || !password) {
+  const { email, password } = req.body;
+  if (!email || !password) {
     return res.status(400).send("Missing data! ");
   }
   let user = User.getUsersByEmail(email);
@@ -78,7 +73,7 @@ router.post("/register", async (req, res) => {
   const salt = await bcrypt.genSalt(12);
   const hashedPwd = await bcrypt.hash(password, salt);
 
-  const savedUser = User.saveUser(name, email, hashedPwd);
+  const savedUser = User.saveUser(email, hashedPwd);
   if (savedUser.changes != 1) {
     return res.status(501).send("User save failed! ");
   }
@@ -87,47 +82,15 @@ router.post("/register", async (req, res) => {
   res.status(200).json(user);
 });
 
-router.put("/:id", async (req, res) => {
-  const { name, email, password } = req.body;
+router.delete("/:id", auth, (req, res) => {
   const id = req.params.id;
-
-  if (!name || !email || !password || !id) {
-    return res.status(400).send("Missing data! ");
+  if (req.userId != id) {
+    return res.status(401).send("Unauthorized");
   }
-
   const user = User.getUsersById(id);
   if (!user) {
     return res.status(404).send("User not found!");
   }
-
-  const salt = await bcrypt.genSalt(12);
-  const hashedPwd = await bcrypt.hash(password, salt);
-
-  const updatedUser = User.updateUser(id, name, email, hashedPwd);
-
-  if (updatedUser.changes != 1) {
-    return res.status(501).send("User update failed! ");
-  }
-  res.status(201).send("Updated");
-});
-
-router.delete("/:id", (req, res) => {
-  const id = req.params.id;
-  const user = User.getUsersById(id);
-  if (!user) {
-    return res.status(404).send("User not found!");
-  }
-
-  const workouts = Workout.getWorkoutByUserId(user.id);
-  workouts.forEach((w) => {
-    const workoutExercises = WorkoutExercise.getWorkoutExerciseByWorkoutId(
-      w.id
-    );
-    workoutExercises.forEach((wE) => {
-      WorkoutExercise.deleteWorkoutExercise(wE.id);
-    });
-    Workout.deleteWorkout(w.id);
-  });
 
   const deletedUser = User.deleteUser(id);
   if (deletedUser.changes != 1) {
